feat: move main to be a file based test
