+++ go.mod
@@ -1,10 +1,10 @@
 module github.com/fzipi/seclang_parser
 
-go 1.20
-
-require github.com/antlr4-go/antlr/v4 v4.13.0
+go 1.21
 
 require (
-	github.com/magefile/mage v1.15.0 // indirect
-	golang.org/x/exp v0.0.0-20230725093048-515e97ebf090 // indirect
+	github.com/antlr4-go/antlr/v4 v4.13.0
+	github.com/magefile/mage v1.15.0
 )
+
+require golang.org/x/exp v0.0.0-20230725093048-515e97ebf090 // indirect
